Replace manual time zone handling with use of the tzlocal module
This solves an issue where we would end up with an empty list of
time zones, throwing an exception as a result.

It also resolves to the user's actual Olson time zone name,
instead of the first of all the possible time zone names that
match the current time zone of the user.

The two helper methods are no longer needed, as we can now use
get_localzone() directly.