Teach Photos service to look for updates to the photo library
The 'changeset' endpoint is used to check for updates to the syncToken,
which means the library has changed. We then clear all the cached photo
albums and assets, which will trigger subsequent accesses to albums()
to fetch fresh album data.

This approach does not try to parse the response of the 'changeset'
endpoint, which in theory contains enough information to do in-place
edits of the cached data, as that is a more fragile approach, and
as long as we don't have a public API to consume those in-place
edits is not worth it. This means more aggressive cache-invalidation,
but we can live with that.

Fixes issue #85